+++ pkg/slo_event_producer/config_test.go
@@ -10,6 +10,7 @@ import (
 )
 
 type configTestCase struct {
+	name           string
 	path           string
 	expectedConfig rulesConfig
 	expectedError  bool
@@ -18,23 +19,28 @@ type configTestCase struct {
 func TestConfig_loadFromFile(t *testing.T) {
 	testCases := []configTestCase{
 		{
+			name: "slo rules file with valid syntax",
 			path: "testdata/slo_rules_valid.yaml.golden",
 			expectedConfig: rulesConfig{Rules: []ruleOptions{
 				{
 					SloMatcher: sloMatcher{Domain: "domain"},
 					FailureConditionsOptions: []operatorOptions{
-						{Operator: "numberHigherThan", Key: "statusCode", Value: "500"},
+						operatorOptions{
+							Operator: "numberHigherThan", Key: "statusCode", Value: "500",
+						},
 					},
 					AdditionalMetadata: stringmap.StringMap{"slo_type": "availability"},
 				}}},
 			expectedError: false,
 		},
 		{
+			name:           "slo_rules file with invalid syntax",
 			path:           "testdata/slo_rules_invalid.yaml.golden",
 			expectedConfig: rulesConfig{},
 			expectedError:  true,
 		},
 		{
+			name:           "invalid path",
 			path:           "?????",
 			expectedConfig: rulesConfig{},
 			expectedError:  true,
@@ -42,12 +48,17 @@ func TestConfig_loadFromFile(t *testing.T) {
 	}
 
 	for _, c := range testCases {
-		var config rulesConfig
-		var _, err = config.loadFromFile(c.path)
-		if c.expectedError {
-			assert.Error(t, err)
-			continue
-		}
-		assert.Equal(t, c.expectedConfig, config, "failed config test for path %s", c.path)
+		t.Run(
+			c.name,
+			func(t *testing.T) {
+				var config rulesConfig
+				var _, err = config.loadFromFile(c.path)
+				if c.expectedError {
+					assert.Error(t, err)
+					return
+				}
+				assert.Equal(t, c.expectedConfig, config, "failed config test for path %s", c.path)
+			},
+		)
 	}
 }